🗺️ Atlas: Extracted Tests from Large Modules
Split the following large "Blob" modules by extracting their inline tests into dedicated `tests.rs` files within new module directories:
- `src/index/adjacency.rs`
- `src/index/incremental_adjacency.rs`
- `src/index/temporal_adjacency.rs`
- `src/index/current.rs`
- `src/storage/checkpoint.rs`
- `src/storage/migration.rs`
- `src/storage/tiered_storage.rs`
